fix: changed the ids from int to long

+++ src/main/java/com/airbnl/managerservice/model/Hotel.java
@@ -10,9 +10,9 @@ import java.beans.Transient;
 @Getter
 @Setter
 public class Hotel {
-    private int id;
+    private long id;
     private String name;
-    private int managerId;
-    private int countryId;
+    private long managerId;
+    private long countryId;
     private String countryName;
 }
